improvement: use existing code to render&get events
